+++ packages/destination-actions/src/destinations/avo/sendSchema/__tests__/snapshot.test.ts
