+++ packages/destination-actions/src/destinations/avo/sendSchema/generated-types.ts
